Refactor component file structure for consistency
Standardize component file organization by moving files into dedicated subdirectories. Adjusted imports across the project to reflect the new structure, ensuring better maintainability and clarity in the codebase.
