+++ src/components/LLMProviderToggle/LLMProviderToggle.tsx
@@ -1,7 +1,7 @@
 // components/LLMProviderToggle.tsx
 import React, {useState} from 'react';
 import {CircleStackIcon} from "@heroicons/react/24/outline";
-import {LLMProvidersModal} from './LLMProvidersModal';
+import {LLMProvidersModal} from "../LLMProviderModal/LLMProvidersModal.tsx";
 
 interface LLMProviderToggleProps {
     selectedProvider: string | null;
